requirements.txt

-e .
albumentations
kornia
matplotlib
pandas
pyyaml
torch
torchmetrics
torchvision
tqdm
umap-learn
wandb
